refactor: rename citation agent to Grounded Document Agent
