+++ ai_agents/grounded_document_agent/assets/how_it_works.png
