+++ ai_agents/grounded_document_agent/grounded_agent/__init__.py
@@ -1,3 +1,3 @@
-"""Citation-aware document Q&A over long PDFs."""
+"""Grounded document Q&A over long PDFs."""
 
 __all__ = ["config", "parser", "indexer", "query", "overview"]
