+++ ai_agents/grounded_document_agent/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "citation-deep-research-agent"
+name = "grounded-document-agent"
 version = "1.0.0"
-description = "Citation-aware document Q&A over long PDFs: LlamaParse for parsing, local Ollama for answers, every answer cited back to the source."
+description = "Grounded document Q&A over long PDFs: LlamaParse for parsing, local Ollama for answers, every answer cited back to the source."
 requires-python = ">=3.10"
 dependencies = [
     "llama-index>=0.12.0",
@@ -17,4 +17,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["citation_agent"]
+packages = ["grounded_agent"]
